Assignment 3 - Ethereum Smart Contract/requirements.txt
absl-py==0.13.0
appdirs==1.4.4
asn1crypto==0.24.0
astor==0.8.1
astunparse==1.6.3
attrdict==2.0.1
attrs==19.3.0
Automat==0.6.0
backcall==0.1.0
beautifulsoup4==4.6.0
binwalk==2.1.1
bleach==1.5.0
blinker==1.4
boto3==1.12.22
botocore==1.15.22
brewer2mpl==1.4.1
cached-property==1.5.2
cachetools==4.2.2
certifi==2021.5.30
chardet==3.0.4
charset-normalizer==2.0.4
click==6.7
cloud-init==19.4
colorama==0.3.7
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
coverage==4.4.1
cryptography==2.1.4
cycler==0.10.0
cytoolz==0.11.0
data==0.4
dataclasses==0.8
decorator==4.4.2
defusedxml==0.6.0
dgl==0.4.2
dgl-cu101==0.4.3.post2
distlib==0.3.0
distro-info===0.18ubuntu0.18.04.1
docutils==0.15.2
entrypoints==0.3
enum34==1.1.10
eth-abi==1.3.0
eth-account==0.3.0
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.2.4
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.10.0
fastdtw==0.3.4
filelock==3.0.12
flake8==3.4.1
flatbuffers==1.12
funcsigs==1.0.2
future==0.18.2
futures==3.1.1
gast==0.4.0
ge==0.0.2
gensim==3.8.1
google-api-core==1.16.0
google-auth==1.34.0
google-auth-oauthlib==0.4.5
google-cloud-core==1.3.0
google-cloud-storage==1.26.0
google-pasta==0.2.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
grpcio==1.34.1
h5py==3.1.0
hexbytes==0.2.2
html5lib==0.9999999
httplib2==0.9.2
hyperlink==17.3.1
idna==3.2
imageio==2.8.0
importlib-metadata==4.6.3
importlib-resources==1.5.0
incremental==16.10.1
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.16.0
Jinja2==2.11.1
jmespath==0.9.5
joblib==0.14.1
jsonpatch==1.16
jsonpointer==1.10
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.0.0
jupyter-console==6.1.0
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.6.3
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
Keras==2.3.1
Keras-Applications==1.0.8
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.1.0
language-selector==0.1
latex==0.7.0
latexbuild==0.2.2
lru-dict==1.1.7
lxml==4.5.0
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.2.0
mccabe==0.6.1
MedPy==0.4.0
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.4
netifaces==0.10.4
networkx==2.4
nibabel==2.2.1
nose==1.3.7
nose-parameterized==0.3.4
notebook==6.0.3
numexpr==2.6.4
numpy==1.19.5
oauthlib==3.1.1
olefile==0.45.1
opencv-python==4.2.0.34
opt-einsum==3.3.0
PAM==0.4.2
pandas==1.0.2
pandocfilters==1.4.2
parsimonious==0.8.1
parso==0.6.2
pdf-to-markdown==0.1.0
pdfminer.six==20200517
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.1
pipenv==2018.11.26
prometheus-client==0.7.1
prompt-toolkit==3.0.4
protobuf==3.17.3
ptyprocess==0.6.0
py==1.4.34
py-solc==3.2.0
py4j==0.10.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.3.1
pycrypto==2.6.1
pycryptodome==3.9.8
pydot==1.2.3
pyflakes==1.5.0
Pygments==2.6.1
PyGObject==3.26.1
PyJWT==1.5.3
PyOpenGL==3.1.0
pyOpenSSL==17.5.0
pyparsing==2.4.6
pyqtgraph==0.10.0
pyrsistent==0.15.7
pyserial==3.4
pyspark==3.1.1
pytest==3.2.2
pytest-cov==2.5.1
python-apt==1.6.5+ubuntu0.2
python-dateutil==2.8.1
python-debian==0.1.32
pytz==2019.3
PyWavelets==1.1.1
pyxdg==0.25
PyYAML==5.3.1
pyzmq==19.0.0
qtconsole==4.7.1
QtPy==1.9.0
requests==2.26.0
requests-oauthlib==1.3.0
requests-unixsocket==0.1.5
rlp==2.0.1
rsa==4.7.2
s3transfer==0.3.3
scikit-image==0.17.2
scikit-learn==0.22.2.post1
scipy==1.4.1
SecretStorage==2.3.1
semantic-version==2.8.5
Send2Trash==1.5.0
service-identity==16.0.0
shutilwhich==1.1.0
SimpleITK==1.2.4
six==1.15.0
smart-open==1.10.0
snap==0.5
snap-stanford==6.0.0
sortedcontainers==2.2.2
ssh-import-id==5.7
stellargraph==0.10.0
svn==1.0.1
systemd-python==234
tables==3.4.2
tempdir==0.7.1
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==1.4.0
tensorflow-estimator==2.5.0
tensorflow-gpu==1.1.0
tensorflow-tensorboard==0.4.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
Theano==0.9.0
threadpoolctl==2.1.0
tifffile==2020.6.3
toolz==0.11.1
torch==1.9.0
torchvision==0.10.0
tornado==6.0.4
tqdm==4.46.0
traitlets==4.3.3
Twisted==17.9.0
typing-extensions==3.7.4.3
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.6
virtualenv==20.0.20
virtualenv-clone==0.5.4
wcwidth==0.1.8
web3==4.9.0
webencodings==0.5.1
websockets==6.0
Werkzeug==2.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xlrd==1.2.0
xlwt==1.3.0
zipp==3.5.0
zope.interface==4.3.2
